# packages/userManagement/swagger.yaml
openapi: 3.0.3
info:
  title: iot-callbot-node
  description: |-

  contact:
    email: nguyenduyphuong3103@gmail.com
  version: 1.0.11
servers:
  - url: http://localhost:8000/api/
...
paths:
...
  /admin/readUsers:
    get:
      tags:
        - admin
      summary: See all Users
      description: "See all Users"
      operationId: readUsers
      security:
        - BearerAuth: []
      parameters:
        - in: query
          name: page
          required: false
          schema:
            type: integer
            minimum: 1
          description: The page number for pagination
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            minimum: 1
          description: The maximum number of users per page
        - in: query
          name: userId
          required: false
          schema:
            type: string
            minimum: 1
          description: The userId for find User
        - in: query
          name: user's name
          required: false
          schema:
            type: string
            minimum: 1
          description: The name of user
      responses:
        default:
          description: found Users
          content:
            application/json:
              schema:
                type: object
                properties:
                  meta:
                    type: object
                    properties:
                      success:
                        type: boolean
                        example: true
                      message:
                        type: string
                        example: "Find Users successfully !!!"
                  resData:
                    type: array
                    items:
                      $ref: "#/components/schemas/ReadUsers"
  /admin/readProfile/{id}:
    get:
      tags:
        - admin
      summary: See user's profile
      description: "See user's profile"
      operationId: readProfile
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: page
          required: true
          schema:
            type: string
            example: e3936164-7caf-480d-949f-7d822ede4a61
          description: Id of user
      responses:
        default:
          description: found information user
          content:
            application/json:
              schema:
                type: object
                properties:
                  meta:
                    type: object
                    properties:
                      success:
                        type: boolean
                        example: true
                      message:
                        type: string
                        example: "Found information successfully !!!"
                  resData:
                    type: array
                    items:
                      $ref: "#/components/schemas/UserRegister"
  /admin/editUserEmail/{id}:
    patch:
      tags:
        - admin
      summary: Edit user email by id
      description: "Edit user email by id"
      operationId: editUserEmail
      security:
        - BearerAuth: []
      parameters:
        - name: userId
          in: path
          description: "id of user"
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  example: "nguyenduyphuong3103@gmail.com"
      responses:
        default:
          description: Edited user by user
          content:
            application/json:
              schema:
                type: object
                properties:
                  meta:
                    type: object
                    properties:
                      ok:
                        type: boolean
                        example: "true"
                      message:
                        type: string
                        example: "The data has been updated!!!"
  /admin/lockUser/{id}:
    patch:
      tags:
        - admin
      summary: Lock user email by id
      description: "Lock user email by id"
      operationId: lockUser
      security:
        - BearerAuth: []
      parameters:
        - name: userId
          in: path
          description: "id of user"
          required: true
          schema:
            type: string
      responses:
        default:
          description: Locked user by user
          content:
            application/json:
              schema:
                type: object
                properties:
                  meta:
                    type: object
                    properties:
                      ok:
                        type: boolean
                        example: "true"
                      message:
                        type: string
                        example: "User has been deactivated !!!"
  /admin/unLockUser/{id}:
    patch:
      tags:
        - admin
      summary: Unlock user email by id
      description: "Unlock user email by id"
      operationId: unLockUser
      security:
        - BearerAuth: []
      parameters:
        - name: userId
          in: path
          description: "id of user"
          required: true
          schema:
            type: string
      responses:
        default:
          description: Unlocked user by user
          content:
            application/json:
              schema:
                type: object
                properties:
                  meta:
                    type: object
                    properties:
                      ok:
                        type: boolean
                        example: "true"
                      message:
                        type: string
                        example: "User has been activated  !!!"
  /admin/createServiceByAdmin:
    post:
      tags:
        - admin
      summary: Create service
      description: "Create service"
      operationId: createServiceByAdmin
      security:
        - BearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceByAdmin"
      responses:
        default:
          description: Created service by admin
          content:
            application/json:
              schema:
                type: object
                properties:
                  meta:
                    type: object
                    properties:
                      ok:
                        type: boolean
                        example: "true"
                      message:
                        type: string
                        example: "Service registration successfully!!!"
                  resData:
                    $ref: "#/components/schemas/ServiceByAdmin"
...
components:
  schemas:
...
    ServiceByAdmin:
      type: object
      properties:
        name:
          type: string
          example: "service 1"
        introduction:
          type: string
          example: "tạo cuộc hội thoại trò chuyện trên skype"
        information:
          type: string
          example: "tạo cuộc hội thoại trò chuyện trên skype"
        price:
          type: number
          example: 1000
        createdAt:
          type: string
          example: "2024-05-22T04:00:50.285Z"
        updatedAt:
          type: string
          example: "2024-05-22T04:00:50.285Z"
      xml:
        name: admin
...
    UserRegister:
      type: object
      properties:
        email:
          type: string
          example: "nguyenduyphuong3103@gmail.com"
        name:
          type: string
          example: "Nguyen Duy Phuong"
        password:
          type: string
          example: "123456"
        passwordConfirm:
          type: string
          example: "123456"
      xml:
        name: user
    ReadUsers:
      type: object
      properties:
        id:
          type: string
          example: "e3936164-7caf-480d-949f-7d822ede4a61"
        name:
          type: string
          example: "Nguyen Duy Phuong"
      xml:
        name: user
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT